Updated headers.  Parital fix for issue #2
Updated code headers to 2018.
Added partial fix for issue #2 - Read6 and Write6 commands to a STOPped LUN will now auto-start the LUN (if the LUN image is available on the SD card).  This does not fix the 'not found' issue.
Debug turned on by default (for development) and reported version number bumped to V002.001.